requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=1.2.1=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiohttp-cors=0.7.0=pypi_0
aiosignal=1.3.2=pypi_0
airportsdata=20241001=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
astor=0.8.1=pypi_0
attr=2.5.1=h166bdaf_1
attrs=24.3.0=pypi_0
av=14.0.1=pypi_0
binutils_impl_linux-64=2.40=h5293946_0
binutils_linux-64=2.40=hb3c18ed_4
blake3=1.0.1=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.12.14=hbcca054_0
cachetools=5.5.0=pypi_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.0=pypi_0
colorama=0.4.6=pypi_0
colorful=0.5.6=pypi_0
compressed-tensors=0.8.1=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
deepspeed=0.16.2=pypi_0
depyf=0.18.0=pypi_0
dill=0.3.9=pypi_0
diskcache=5.6.3=pypi_0
distlib=0.3.9=pypi_0
distro=1.9.0=pypi_0
einops=0.8.0=pypi_0
expat=2.6.4=h6a678d5_0
fastapi=0.115.6=pypi_0
filelock=3.16.1=pypi_0
flask=3.1.0=pypi_0
fonttools=4.55.3=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.2.0=pypi_0
gcc=11.4.0=h602e360_13
gcc_impl_linux-64=11.4.0=h00c12a0_13
gcc_linux-64=11.4.0=ha077dfb_4
gguf=0.10.0=pypi_0
google-api-core=2.24.0=pypi_0
google-auth=2.37.0=pypi_0
googleapis-common-protos=1.66.0=pypi_0
grpcio=1.69.0=pypi_0
gxx_impl_linux-64=11.4.0=h634f3ee_13
gxx_linux-64=11.4.0=h35bfe5d_4
h11=0.14.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.27.0=pypi_0
idna=3.10=pypi_0
imageio=2.36.1=pypi_0
importlib-metadata=8.5.0=pypi_0
iniconfig=2.0.0=pypi_0
interegular=0.3.3=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.3=pypi_0
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
kernel-headers_linux-64=3.10.0=he073ed8_18
kiwisolver=1.4.8=pypi_0
lark=1.2.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libcap=2.71=h39aace5_0
libevent=2.1.12=hf998b51_1
libexpat=2.6.4=h5888daf_0
libfabric=1.22.0=ha770c72_3
libfabric1=1.22.0=ha594dbc_3
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-devel_linux-64=11.4.0=h8f596e0_113
libgcc-ng=14.2.0=h69a702a_1
libgcrypt-lib=1.11.0=hb9d3cd8_2
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgomp=14.2.0=h77fa898_1
libgpg-error=1.51=hbd13f7d_1
libhwloc=2.11.2=default_h0d58e46_1001
libiconv=1.17=hd590300_2
liblzma=5.6.3=hb9d3cd8_1
libnl=3.11.0=hb9d3cd8_0
libnsl=2.0.1=hd590300_0
libsanitizer=11.4.0=h5763a12_13
libsqlite=3.45.2=h2797004_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-devel_linux-64=11.4.0=h8f596e0_113
libstdcxx-ng=14.2.0=h4852527_1
libsystemd0=256.9=h0b6a36f_2
libudev1=256.9=h9a4d06a_2
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.13.5=h0d44e9d_1
libzlib=1.3.1=hb9d3cd8_2
linkify-it-py=2.0.3=pypi_0
lm-format-enforcer=0.10.9=pypi_0
lz4-c=1.10.0=h5888daf_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.0=pypi_0
mdit-py-plugins=0.4.2=pypi_0
mdurl=0.1.2=pypi_0
memray=1.15.0=pypi_0
mistral-common=1.5.1=pypi_0
modelscope=1.22.0=pypi_0
mpi=1.0=openmpi
mpi4py=4.0.1=py312h5ca6011_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
ninja=1.11.1.3=pypi_0
nltk=3.9.1=pypi_0
numpy=1.26.3=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-ml-py=12.560.30=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
openai=1.59.4=pypi_0
opencensus=0.11.4=pypi_0
opencensus-context=0.1.3=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openmpi=5.0.6=hd45feaf_100
openssl=3.4.0=hb9d3cd8_0
outlines=0.1.11=pypi_0
outlines-core=0.1.26=pypi_0
packaging=24.2=pypi_0
partial-json-parser=0.2.1.1.post4=pypi_0
peft=0.14.0=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py312h06a4308_0
platformdirs=4.3.6=pypi_0
pluggy=1.5.0=pypi_0
prometheus-client=0.21.1=pypi_0
prometheus-fastapi-instrumentator=7.0.0=pypi_0
propcache=0.2.1=pypi_0
proto-plus=1.25.0=pypi_0
protobuf=5.29.2=pypi_0
psutil=6.1.1=pypi_0
py-cpuinfo=9.0.0=pypi_0
py-spy=0.4.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pybind11=2.13.6=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.8=pypi_0
pycountry=24.6.1=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.1=pypi_0
pyte=0.8.2=pypi_0
pytest=8.3.4=pypi_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python_abi=3.12=5_cp312
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
qwen-vl-utils=0.0.8=pypi_0
ray=2.40.0=pypi_0
rdma-core=55.0=h5888daf_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.5=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.1.0=py312h06a4308_0
six=1.17.0=pypi_0
smart-open=7.1.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.2=h2c6b66d_0
starlette=0.41.3=pypi_0
sympy=1.13.1=pypi_0
sysroot_linux-64=2.17=h0157908_18
textual=1.0.0=pypi_0
thefuck=3.32=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.21.0=pypi_0
torch=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.47.1=pypi_0
triton=3.1.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024b=h04d1e81_0
uc-micro-py=1.0.3=pypi_0
ucc=1.3.0=h0f835a6_3
ucx=1.17.0=h53fb5aa_4
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
virtualenv=20.28.1=pypi_0
vllm=0.6.6.post1=pypi_0
watchfiles=1.0.3=pypi_0
wcwidth=0.2.13=pypi_0
websockets=14.1=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py312h06a4308_0
wrapt=1.17.0=pypi_0
xformers=0.0.28.post3=pypi_0
xgrammar=0.1.9=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.18.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.3.1=hb9d3cd8_2
zstd=1.5.6=ha6fb4c9_0
